+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Fixed
+- Possible crash in PlayerDataHandler [#392]
+
 ## [2.8.21] - 2019-07-31
 ### Fixed
 - Server crash